// src/modules/reports/dto/report-interfaces.ts
export interface ITabularField {
  id: string;
  draggedId: string;
  columnName?: string;
  columnDisplayName: string;
  type: string;
  operation?: string;
  node?: string;
  isExplodedBy?: boolean;
}

export interface IReportField extends ITabularField {
  footerAggregation: Array<string>;
  decimalNumbers?: number;
  thresholdUpperValue?: number;
  thresholdLowerValue?: number;
  dateFormat?: string;
  hidden: boolean;
  pinned: boolean;
  trunc?: boolean;
  round?: boolean;
  index: number;
  trValue?: number;
  selected?: boolean;
  isMetric?: boolean;
}

export interface IMinimalTabularTable {
  id: string;
  displayName: string;
  role?: string;
  fields: Array<IReportField>;
}
// ...
export interface IWhenCondition {
  whenStatementId: string;
  fieldStatementId: string;
  whenStatementType: string;
  whenStatement: string;
  operator: string;
  fieldStatement: string;
  thenStatementId: string;
  thenStatement: string;
  isThenCustomField: boolean;
  isCustomField: boolean;
  isSubCondition: boolean;
  isFieldCustomColumn: boolean;
  isThenCustomColumn: boolean;
  isWhenCustomColumn: boolean;
  condition: string;
  subCondition: IWhenSubCondition;
}

export interface IControlValue {
  elseStatement: string;
  elseStatementId: string;
  elseStatementTableId: string;
  iselseCustomField: boolean;
  whenCondition: Array<IWhenCondition>;
  isElseCustomColumn: boolean;
}

export interface IOperationColumn {
  isOperation: boolean;
  isCustomColumn: boolean;
  draggedId: string;
}

export interface IOperationOperatorColumn {
  isOperation: boolean;
  operator: string;
}

export type OperationColumnType = IOperationOperatorColumn | IOperationColumn;

export interface ICustomControlColumn {
  draggedId: string;
  columnDisplayName: string;
  footerAggregation: Array<string>;
  type: string;
  controlValue: IControlValue;
  customColumnType: string;
  customColumn: boolean;
  operation: string;
  hidden: boolean;
  pinned: boolean;
  trunc: boolean;
  round: boolean;
  index: number;
  trValue: number;
  selected?: boolean;
  isMetric?: boolean;
  isExplodedBy?: boolean;
}

export interface ICustomOperationColumn {
  draggedId: string;
  columnDisplayName: string;
  footerAggregation: Array<string>;
  type: string;
  operation: string;
  savedTokens: Array<OperationColumnType>;
  customColumn: boolean;
  customColumnType: string;
  hidden: boolean;
  pinned: boolean;
  index: number;
  trunc: boolean;
  round: boolean;
  trValue: number;
  selected?: boolean;
  isMetric?: boolean;
  isExplodedBy?: boolean;
}

export interface ICustomCompareColumn {
  draggedId: string;
  columnDisplayName: string;
  timeFilter: string;
  backPeriod: number;
  customColumn: boolean;
  customColumnType: string;
  operation: string;
  footerAggregation: Array<string>;
  withStatDate: boolean;
  dateFormat: string;
  hidden: boolean;
  pinned: boolean;
  isCustom: boolean;
  customString?: string;
  tablesUsed?: Array<string>;
  type: string;
  usedColumnId: string;
  trunc: boolean;
  round: boolean;
  index: number;
  trValue: number;
  selected?: boolean;
  isMetric?: boolean;
  isExplodedBy?: boolean;
}
// ...
export interface IChartData {
  id: string;
  name: string;
  type: string;
  orderIndex: number;
  [key: string]: unknown;
}